+++ src/noggit/World.inl
@@ -15,9 +15,9 @@ void World::for_all_chunks_on_tile (glm::vec3 const& pos, Fun&& fun)
   {
     mapIndex.setChanged(tile);
 
-    for (size_t ty = 0; ty < 16; ++ty)
+    for (unsigned ty = 0; ty < 16; ++ty)
     {
-      for (size_t tx = 0; tx < 16; ++tx)
+      for (unsigned tx = 0; tx < 16; ++tx)
       {
         fun(tile->getChunk(ty, tx));
       }
